% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bm.R
\name{create_bm}
\alias{create_bm}
\title{Create base map}
\usage{
create_bm(
  DT,
  lat = "lat",
  lon = "lon",
  buffer = 1000,
  sc_dist,
  projection = paste0("+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0",
    " +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 ")
)
}
\arguments{
\item{DT}{Name of the data.table}

\item{lat}{Name of the column with latitude (as.character)}

\item{lon}{Name of the column with longitude (as.character)}

\item{buffer}{Buffer around the data}

\item{sc_dist}{Distance of the scale}

\item{projection}{Projection of the data (default is equal area with centre Barrow)}
}
\value{
bm, a ggplot2 base map
}
\description{
Use open street map data to create a basemap with the extent of the data
}
\examples{
# create table with two points
DT = data.table(name = c('NARL', 'Utqiagvik'),
                lat  = c(71.320854, 71.290246),
                lon  = c(-156.648210, -156.788622))

# change projection
st_transform_DT(DT)

# create base map
bm = create_bm(DT)
bm
}
