% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_colored_bm.R
\name{create_colored_bm}
\alias{create_colored_bm}
\title{Create colored base map}
\usage{
create_colored_bm(
  DT,
  lat = "lat",
  lon = "lon",
  buffer = 1000,
  aspect = "16:9",
  sc_dist,
  sc_location = "br",
  sc_cex = 0.7,
  sc_height = unit(0.25, "cm"),
  projection = paste0("+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0",
    " +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 ")
)
}
\arguments{
\item{DT}{Name of the data.table}

\item{lat}{Name of the column with latitude (as.character)}

\item{lon}{Name of the column with longitude (as.character)}

\item{buffer}{Buffer around the data}

\item{aspect}{Set aspect ratio of the base map}

\item{sc_dist}{Distance of the scale}

\item{sc_location}{Where to put the scale bar ("tl" for top left, etc.)}

\item{sc_cex}{Label size as numeric character expansion factor}

\item{sc_height}{Thickness of the scale bar}

\item{projection}{Projection of the data (default is equal area with centre Barrow)}
}
\value{
bm, a ggplot2 base map
}
\description{
Use open street map data to create a basemap with the extent of the data
}
\examples{
# create table with two points
DT = data.table(name = c('NARL', 'Utqiagvik'),
                lat  = c(71.320854, 71.290246),
                lon  = c(-156.648210, -156.788622))

# change projection
st_transform_DT(DT)

# create base map
bm = create_colored_bm(DT)
bm
}
