% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_id_check.R
\name{obs_id_check}
\alias{obs_id_check}
\title{Get names of GPS points before and after}
\usage{
obs_id_check(obs_id, start = -3, end = 3)
}
\arguments{
\item{obs_id}{Observation ID in format Year_GPSID_GPSWaypoint}

\item{start}{N of positions before}

\item{end}{N of positions after}
}
\value{
obs_id's to subset GPS points with
}
\description{
Used to check if the point was realistic
}
\examples{
obs_id_check('18_2_221')
}
