% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_map_data.R
\docType{data}
\name{osm_lakes}
\alias{osm_lakes}
\title{Base map polygons}
\format{
sf polygons from open street map  in '+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 '
\itemize{
  \item lakes
  }
}
\source{
\url{http://download.geofabrik.de/north-america/us/alaska.html}
}
\usage{
data(osm_lakes)
}
\description{
Open street map polygons of lakes
}
\examples{
osm_lakes

}
\references{
\url{http://openstreetmapdata.com}
}
\keyword{datasets}
