% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_land.R
\docType{data}
\name{osm_land}
\alias{osm_land}
\title{Base map polygons}
\format{sf polygons from open street map  in '+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 '
\itemize{
  \item land
  }}
\source{
\url{http://openstreetmapdata.com/data/land-polygons}
}
\usage{
data(osm_land)
}
\description{
Open street map polygons of land
}
\examples{
osm_land

}
\references{
\url{http://openstreetmapdata.com}
}
\keyword{datasets}
