% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_over_poly_DT.R
\name{point_over_poly_DT}
\alias{point_over_poly_DT}
\title{Check if points are over a polygon (e.g. study site)}
\usage{
point_over_poly_DT(
  DT,
  lat = "lat",
  lon = "lon",
  poly,
  buffer = NA,
 
    projection = "+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 "
)
}
\arguments{
\item{DT}{Name of the data.table}

\item{lat}{Name of the column with latitude (as.character)}

\item{lon}{Name of the column with longitude (as.character)}

\item{poly}{Name of the polygon to overlap the data with (changes projection to PROJ)}

\item{buffer}{Buffer distance (see st_buffer)}

\item{projection}{Projection of the data (default is equal area with centre Barrow)}
}
\value{
Data.table with additional conditional column poly_overlap
}
\description{
Check if points are over a polygon (e.g. study site)
}
\examples{
require(data.table)

# load study site
study_site

# create table with one inside and one outside the study site
DT = data.table(name = c('NARL', 'Utqiagvik'),
                lat  = c(71.320854, 71.290246),
                lon  = c(-156.648210, -156.788622))

# change projection
st_transform_DT(DT)
point_over_poly_DT(DT, lat = 'lat', lon = 'lon', poly = study_site)

\dontrun{
# check data
require(ggplot2)
ggplot() +
  geom_sf(data = study_site, fill = 'grey95') +
  geom_point(data = DT, aes(lon, lat, color = poly_overlap))
}
}
