% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_seen.R
\name{prop_seen}
\alias{prop_seen}
\title{Calculate proportion of individuals seen for specific time}
\usage{
prop_seen(DT, category, ID_unique, time_seen)
}
\arguments{
\item{DT}{Data.table with required data}

\item{category}{Column that inculdes categories for which one wants to calculate the proportion seen (as.character)}

\item{ID_unique}{Column that inculdes Unique ID by which one wants to calculte the days seen (as.character)}

\item{time_seen}{Column that inculdes time seen (as.character)}
}
\value{
Data.table with category, time_step and proportion seen
}
\description{
Calculate proportion of individuals seen for specific time
}
\examples{

dt = data.table(category  = c(rep('M', 100), rep('F', 100)),
                ID_unique = as.character(c(1:200)),
                time_seen = sample(x = c(1:40), size = 200, replace = TRUE) )

ds = prop_seen(dt, 'category', 'ID_unique', 'time_seen')

ggplot(data = ds, aes(x = time_step, y = prop_seen, col = category)) +
  scale_y_continuous(limits = c(0, 1)) +
  geom_line(size = 1.5) +
  scale_color_brewer(name = 'Category', palette = 'Dark2') +
  labs(x = 'Days seen', y = 'Proportion of individuals') +
  theme_classic(base_size = 24)
}
