% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform_DT.R
\name{st_transform_DT}
\alias{st_transform_DT}
\title{Change the projection within a data table}
\usage{
st_transform_DT(
  DT,
  lat = "lat",
  lon = "lon",
  from = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
 
    to = "+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 "
)
}
\arguments{
\item{DT}{Name of the data.table}

\item{lat}{Name of the column with latitude (as.character)}

\item{lon}{Name of the column with longitude (as.character)}

\item{from}{Projection of the data (default is longlat)}

\item{to}{Projection that the data should get (default is equal area with centre Barrow)}
}
\value{
Data.table as it was with changed projection
}
\description{
Change the projection within a data table
}
\examples{
require(data.table)
require(sf)

DT = data.table(name = 'Utqiagvik',
                lat  = 71.320854,
                lon  = -156.648210)

st_transform_DT(DT)
}
