% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_site.R
\docType{data}
\name{study_site}
\alias{study_site}
\title{Study site (mainly NARL in Utqiagvik)}
\format{A sf polygon of the study site in '+proj=laea +lat_0=90 +lon_0=-156.653428 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 '
\itemize{
  \item edge: polygon of the study site
  }}
\usage{
data(study_site)
}
\description{
Red Phalarope study site (mainly NARL in Utqiagvik)
}
\examples{
study_site

}
\references{

}
\keyword{datasets}
